requirements.txt
pymodbus>=3.5.0
pyserial>=3.5
